#1596: elm: move to top level to generalize
