+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(TOP_LEVEL_SUBDIRS
   configs
   collective
+  elm
   epoch
   group
   messaging
@@ -33,6 +34,7 @@ set(
       collective/scatter/
       collective/reduce/
         collective/reduce/operators collective/reduce/functors
+    elm
     group/id group/region group/global group/msg group/collective group/rooted
     group/base
     pipe/id pipe/msg pipe/state pipe/signal pipe/interface pipe/callback
